Objective

We provide the following methods to provide a compositional approach for defining common objectives used in joint gas-power models. These methods should always be defined over AbstractGasPowerModel.

GasPowerModels.objective_expression_ne_compressor_cost — Method

Helper function for expressing compressor costs.

GasPowerModels.objective_expression_ne_line_cost — Method

Helper function for expressing line costs.

GasPowerModels.objective_expression_ne_pipe_cost — Method

Helper function for expressing pipe costs.

GasPowerModels.objective_expression_opf_cost — Method

Helper function for constructing the expression associated with the OPF objective.

GasPowerModels.objective_expression_pressure_penalty — Method

Helper function for expressing zonal pressure penalty prices.

GasPowerModels.objective_expression_zone_price — Method

Helper function for expressing zonal prices.

GasPowerModels.objective_max_gas_load — Method

Maximizes the normalized sum of nongeneration gas load delivered in the joint network, i.e.,

\[\max \eta_{G}(d) := \left(\sum_{i \in \mathcal{D}^{\prime}} \beta_{i} d_{i}\right) \left(\sum_{i \in \mathcal{D}^{\prime}} \beta_{i} \overline{d}_{i}\right)^{-1},\]

where $\mathcal{D}^{\prime}$ is the set the delivery points in the gas network with dispatchable demand that are not connected to interdependent generators in the power network, $\beta_{i} \in \mathbb{R}_{+}$ (equal to the priority property of the delivery) is a predefined restoration priority for delivery $i \in \mathcal{D}^{\prime}$, $d_{i}$ is the mass flow of gas delivered at $i \in \mathcal{D}^{\prime}$, and $\overline{d}_{i}$ is the maximum deliverable gas load at $i \in \mathcal{D}^{\prime}$.

GasPowerModels.objective_max_load — Method

Maximizes the weighted normalized sums of nongeneration gas load and active power load delivered in the joint network, i.e.,

\[ \max \lambda_{G} \eta_{G}(d) + \lambda_{P} \eta_{P}(z^{d}),\]

where it is recommended that $0 < \lambda_{G} < 1$, that gm_load_priority in the network data specification be set to the value of $\lambda_{G}$ desired, and that pm_load_priority similarly be set to the value $1 - \lambda_{G} = \lambda_{P}$. This type of parameterization allows for a straightforward analysis of gas-power tradeoffs, as the objective is naturally scaled between zero and one.

GasPowerModels.objective_max_power_load — Method

Maximizes the normalized sum of active power load delivered in the joint network, i.e.,

\[\max \eta_{P}(z^{d}) := \left(\sum_{i \in \mathcal{L}} \beta_{i} z_{i}^{d} \lvert \Re({S}_{i}^{d}) \rvert \right) \left(\sum_{i \in \mathcal{L}} \beta_{i} \lvert \Re({S}_{i}^{d}) \rvert \right)^{-1}.\]

Here, $\mathcal{L}$ is the set of loads in the power network, $\beta_{i} \in \mathbb{R}_{+}$ (equal to the weight property of the load) is the load restoration priority for load $i \in \mathcal{L}$, and $z_{i} \in [0, 1]$ is a variable that scales the absolute maximum amount of active power load, $\lvert \Re({S}_{i}^{d}) \rvert$, at load $i \in \mathcal{L}$.

GasPowerModels.objective_min_ne_cost — Method

Objective for minimizing the costs of expansion. Formally stated as

\[\min \alpha \sum_{(i,j) \in Pipe \cup Compressors} \kappa_{ij} z_{ij} + \beta \sum_{(i,j) \in Branches} \kappa_{ij} z_{ij},\]

where $\alpha$ and $\beta$ are weighting terms.

GasPowerModels.objective_min_ne_opf_cost — Method

Objective function for minimizing the gas-grid optimal flow combined with network expansion costs as defined in reference Russell Bent, Seth Blumsack, Pascal Van Hentenryck, Conrado Borraz-Sánchez, Mehdi Shahriari. Joint Electricity and Natural Gas Transmission Planning With Endogenous Market Feedbacks. IEEE Transactions on Power Systems. 33 (6): 6397-6409, 2018. More formally, this objective is stated as

\[\min \alpha \sum_{(i,j) \in Pipe \cup Compressors} \kappa_{ij} z_{ij} + \beta \sum_{(i,j) \in Branches} \kappa_{ij} z_{ij} + \lambda \sum_{g \in G} (c^1_g pg_g^2 + c^2_g pg_g + c^3_g) + \gamma \sum_{z \in Z} cost_z + \gamma \sum_{z \in Z} pc_z,\]

where $\lambda, \alpha, \beta$ and $\gamma$ are weighting terms.

GasPowerModels.objective_min_opf_cost — Method

Objective function for minimizing the gas-grid optimal flow as defined in reference Russell Bent, Seth Blumsack, Pascal Van Hentenryck, Conrado Borraz-Sánchez, Mehdi Shahriari. Joint Electricity and Natural Gas Transmission Planning With Endogenous Market Feedbacks. IEEE Transactions on Power Systems. 33 (6): 6397-6409, 2018. More formally, this objective is stated as

\[\min \lambda \sum_{g \in G} (c^1_g pg_g^2 + c^2_g pg_g + c^3_g) + \gamma \sum_{z \in Z} cost_z + \gamma \sum_{z \in Z} pc_z,\]

where $\lambda$ and $\gamma$ are weighting terms.