Variables

We provide the following methods to provide a compositional approach for defining common variables used in coupled gas grid flow models. These methods should always be defined over AbstractGasModel and/or AbstractPowerModel.

GasPowerModels.getstart — Function

extracts the start value

GasPowerModels.variable_pressure_price — Function

Initializes variables associated with zonal pressure cost.

GasPowerModels.variable_zone_demand — Function

function for creating variables associated with zonal demand

GasPowerModels.variable_zone_demand_price — Function

function for creating variables associated with zonal demand

GasPowerModels.variable_zone_pressure — Function

Initializes variables associated with zonal demand.