Variables

Variables

We provide the following methods to provide a compositional approach for defining common variables used in coupled gas grid flow models. These methods should always be defined over "GenericGasModels" and/or "GenericPowerModels".

GasPowerModels.getstart — Function.

extracts the start value

source

GasPowerModels.variable_pressure_price — Function.

function for creating variables associated with zonal pressure cost

source

GasPowerModels.variable_zone_demand — Function.

function for creating variables associated with zonal demand

source

GasPowerModels.variable_zone_demand_price — Function.

function for creating variables associated with zonal demand

source

GasPowerModels.variable_zone_pressure — Function.

function for creating variables associated with zonal demand

source